feat(Winner) : 우승자 출력 역할 분리
