+++ docs/README.md
@@ -25,3 +25,10 @@
 [x] 레이스
 - 자동차 객체 && 이동 객체 활용
 - 레이스 실행 && 레이스 결과 출력
+
+[x] 우승자 출력
+- 레이스 결과 활용하여 최종 우승자 return
+- 복수 가능
+- 레이스 결과 출력을 자동차 수만큼 반복
+- 최종 우승자 출력
+
