+++ src/main/java/racingcar/Application.java
@@ -1,11 +1,5 @@
 package racingcar;
 
-import camp.nextstep.edu.missionutils.Console;
-import camp.nextstep.edu.missionutils.Randoms;
-
-
-
-
 public class Application {
     public static void main(String[] args) {
 
@@ -24,42 +18,6 @@ public class Application {
             Race.excute(cars, positions);
         }
 
-        printWinners(cars, positions);
-    }
-
-
-
-
-
-
-
-    private static void printWinners(String[] carNames, int[] positions) {
-        int maxPosition = 0;
-        for (int position : positions) {
-            maxPosition = getMaxPosition(maxPosition, position);
-        }
-
-        StringBuilder winners = new StringBuilder();
-        for (int i = 0; i < carNames.length; i++) {
-            if (positions[i] == maxPosition) {
-                appendWinners(winners, carNames[i]);
-            }
-        }
-
-        System.out.println("최종 우승자 : " + winners.toString());
-    }
-
-    private static int getMaxPosition(int maxPosition, int position) {
-        if (position > maxPosition) {
-            maxPosition = position;
-        }
-        return maxPosition;
-    }
-
-    private static void appendWinners(StringBuilder winners, String carNames) {
-        if (winners.length() > 0) {
-            winners.append(", ");
-        }
-        winners.append(carNames);
+        Winner.choose(cars, positions);
     }
 }
